requirements.txt
albumentations
opencv-python
Pillow
numpy
segmentation-models-pytorch
torch
torchvision
tqdm